refactor: 검증 로직 Coupon 내부로 이동
